# Faster wallet refreshes