@glv2 2018-05-21
Here are some test results of a wallet scanning the whole blockchain:

computer
 - OS: GNU/Linux x86-64
 - CPU: Intel Core i7-3630QM @ 2.40GHz
 - Storage: Samsung SSD 850 PRO

master branch (4b728d7dd48584987f53995a141baac4f886f017)
 - blocks: 1577488
 - time: 31.3 minutes

master branch + PR3716
 - blocks: 1577521
 - time: 8.6 minutes

master branch + PR3716 + refresh-type=no-coinbase
 - blocks: 1577526
 - time: 7.2 minutes