@moneromooo-monero 2018-06-10
Now ready for review.