@stoffu 2018-06-11
Computer:

- OS: Windows 7 Professional, 64bit
- CPU: Intel Core i7-5690X @ 3.00GHz
- Storage: Samsung SSD 850 PRO

Master (25e7a7d9):
- optimize-coinbase -> 26m 46s (1592555 blocks)
- no-coinbase -> 24m 24s (1592573 blocks)

This patch (c4fda6d0):
- optimize-coinbase -> 5m 31s (1592557 blocks)
- no-coinbase -> 4m 52s (1592578 blocks)

I can do two more tests on OSX machines with SSD and HDD.