@stoffu 2018-06-11
Note for testers: if you want to measure the time of scanning the entire blockchain from the genesis, not only the restore height has to be set to 0, but also the wallet creation time must not be set to current. That is, if you create a new wallet and `set refresh-from-block-height 0` and then `rescan_bc`, the wallet will NOT do the full scan of the blockchain (instead it skips old blocks based on timestamps). The correct way is to restore the wallet from seed and setting the restore height to 0.