@stoffu 2018-06-11
Computer:
- OS: Mac OSX 10.12.6
- CPU: 2.9 GHz Intel Core i7
- Storage: APPLE SSD SM2048L

Master:
- optimize-coinbase -> 23m 12s (1592682 blocks)
- no-coinbase -> 21m 31s (1592694 blocks)

PR-3716:
- optimize-coinbase -> 11m 9s (1592697 blocks)
- no-coinbase -> 9m 17s (1592702 blocks)