@stoffu 2018-06-13
Computer:
- OS: macOS Sierra 10.12.6
- CPU: 2.3 GHz Intel Core i7
- Storage: APPLE HDD HTS541010A9E662

Master:
- optimize-coinbase -> 1h 58m 24s (1592635 blocks)
- no-coinbase -> 1h 56m 42s (1593176 blocks)

PR-3716:
- optimize-coinbase -> 2h 0m 18s (1593251 blocks)
- no-coinbase -> 1h 57m 50s (1593310 blocks)