@SamsungGalaxyPlayer 2018-06-14
I will test this on the new 970 EVO when I get a chance. I'll update this comment when done.